* commit '42c5517ec8502fdf3442f2fc0c5d53c99f44d2ad':
  Iterator
  BI BUG
  不使用Field的默认构造
  更新进度，version
  bug

+++ fbi_engine_base/src/com/fr/bi/cal/stable/tableindex/AbstractTableIndex.java
@@ -136,7 +136,7 @@ public abstract class AbstractTableIndex implements ICubeTableService {
     }
 
     @Override
-    public int getTableVersion(BIKey key) {
+    public long getTableVersion(BIKey key) {
         return tableVersion;
     }
